move uv calculations to polygon mesh modifier remove uv modifier class
