--- sdkproject/Assets/Mapbox/Examples/1_DataExplorer/Modules/ExplorerUvModifier.asset
@@ -1,15 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: cf1a49b7a5d29964093712def3d56475, type: 3}
-  m_Name: ExplorerUvModifier
-  m_EditorClassIdentifier: 
-  Active: 1
-  UseSatelliteRoof: 0
