--- sdkproject/Assets/Mapbox/Examples/1_DataExplorer/Modules/ExplorerUvModifier.asset.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 23fec0796c24f584ca6c13522eb63102
-timeCreated: 1511897673
-licenseType: Pro
-NativeFormatImporter:
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
