--- sdkproject/Assets/Mapbox/Examples/_sharedModules/UvModifier.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b432bf85b0df280468dcfdaf5a9b90d0
-timeCreated: 1485209375
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
