+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/LayerVisualizers/VectorLayerVisualizer.cs
@@ -254,15 +254,15 @@ namespace Mapbox.Unity.MeshGeneration.Interfaces
 							AddOrCreateMeshModifier<SnapTerrainModifier>();
 						}
 
-						AddOrCreateMeshModifier<PolygonMeshModifier>();
+						var poly = AddOrCreateMeshModifier<PolygonMeshModifier>();
 
 						UVModifierOptions uvModOptions = new UVModifierOptions();
 						uvModOptions.texturingType = (_layerProperties.materialOptions.style == StyleTypes.Custom) ? _layerProperties.materialOptions.customStyleOptions.texturingType : _layerProperties.materialOptions.texturingType;
 						uvModOptions.atlasInfo = (_layerProperties.materialOptions.style == StyleTypes.Custom) ? _layerProperties.materialOptions.customStyleOptions.atlasInfo : _layerProperties.materialOptions.atlasInfo;
 						uvModOptions.style = _layerProperties.materialOptions.style;
-
-						var uvMod = AddOrCreateMeshModifier<UvModifier>();
-						uvMod.SetProperties(uvModOptions);
+						poly.SetProperties(uvModOptions);
+						//var uvMod = AddOrCreateMeshModifier<UvModifier>();
+						//uvMod.SetProperties(uvModOptions);
 
 						if (_layerProperties.extrusionOptions.extrusionType != Map.ExtrusionType.None)
 						{
