+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Modifiers/GameObjectModifiers/MaterialModifier.cs
@@ -48,7 +48,7 @@ namespace Mapbox.Unity.MeshGeneration.Modifiers
 			{
 				for (int i = 0; i < min; i++)
 				{
-					mats[i] = _options.materials[i].Materials[UnityEngine.Random.Range(0, _options.materials[i].Materials.Length)];
+					mats[i] = Instantiate(_options.materials[i].Materials[UnityEngine.Random.Range(0, _options.materials[i].Materials.Length)]);
 				}
 
 				mats[0].mainTexture = tile.GetRasterData();
