--- sdkproject/Assets/Mapbox/Unity/MeshGeneration/Modifiers/MeshModifiers/UvModifier.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: cf1a49b7a5d29964093712def3d56475
-timeCreated: 1478553093
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
