+++ sdkproject/Assets/Mapbox/Unity/Tests/ApiTest.cs
@@ -401,7 +401,6 @@ public class ApiTest : MonoBehaviour
 			new List<MeshModifier>()
 			{
 				ScriptableObject.CreateInstance<PolygonMeshModifier>(),
-				ScriptableObject.CreateInstance<UvModifier>(),
 				ScriptableObject.CreateInstance<HeightModifier>()
 			},
 			new List<GameObjectModifier>()
