--- sdkproject/Assets/MapboxAR/Examples/ARTabletop/ARTableMapVisualizer/ClipBuildingUvModifier.asset.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 239f012c0bd1044da97461c3374ed178
-timeCreated: 1502384544
-licenseType: Pro
-NativeFormatImporter:
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
