--- sdkproject/Assets/MapboxAR/Examples/SimpleMap/SimpleBuildingUvModifier.asset.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 78682c836b14e46c0b705180848dc333
-timeCreated: 1502384544
-licenseType: Pro
-NativeFormatImporter:
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
